requirements.txt
numpy==1.18.3
pyro-api==0.1.1
pyro-ppl==1.3.1
torch==1.5.0
torchtext==0.6.0
torchvision==0.6.0
tqdm==4.45.0
